requirements.txt
telethon>=1.34.0
python-dotenv>=1.0.0
openai>=1.30.0
apscheduler>=3.10.0
aiohttp>=3.9.0
aiohttp-jinja2>=1.6.0
jinja2>=3.1.0
markdown>=3.6
